requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
async-timeout==5.0.1
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
cryptography==46.0.6
dnspython==2.7.0
ecdsa==0.19.2
email-validator==2.3.0
Events==0.5
exceptiongroup==1.3.1
fastapi==0.128.8
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
MarkupSafe==3.0.3
opensearch-py==3.0.0
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-exporter-otlp-proto-http==1.41.0
opentelemetry-instrumentation==0.62b0
opentelemetry-instrumentation-asgi==0.62b0
opentelemetry-instrumentation-fastapi==0.62b0
opentelemetry-instrumentation-redis==0.62b0
opentelemetry-instrumentation-requests==0.62b0
opentelemetry-instrumentation-sqlalchemy==0.62b0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opentelemetry-util-http==0.62b0
packaging==26.0
passlib==1.7.4
protobuf==6.33.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
redis==7.0.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==0.49.3
stripe==15.0.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn==0.39.0
wrapt==2.1.2
zipp==3.23.0
